Add CXXType to its own header
